Bellabeat - Fitness Data
This code was used for the Google Data Analytics Certification - Capstone Project. The project was based on the Bellabeat products, it was one of two guided projects within the course.

The task:  Analyze fitbit data to provide the marketing team with valuable insights on how to better market their products.